Move Access component to VisibilityStep. Verified that API receives the machine request with updated access list
