+++ troposphere/static/js/components/modals/instance/image/steps/VisibilityStep.jsx
@@ -2,12 +2,14 @@ import React from "react";
 import RaisedButton from "material-ui/RaisedButton";
 import Backbone from "backbone";
 import Visibility from "../components/Visibility";
+import EditAccessView from "components/images/detail/access_list/EditAccessView";
 import Users from "components/common/Users";
 
 export default React.createClass({
     displayName: "ImageWizard-VisibilityStep",
 
     propTypes: {
+        allPatterns: React.PropTypes.instanceOf(Backbone.Collection).isRequired,
         instance: React.PropTypes.instanceOf(Backbone.Model).isRequired
     },
 
@@ -19,13 +21,25 @@ export default React.createClass({
     },
 
     getInitialState: function() {
+        // let parentImage = stores.ImageStore.get(instance.get('image').uuid);
+        //Note: An improvement here would be showing a `loader` until ImageStore.get() returns
+        // and then bootstrapping `activeAccessList` with `image.get('access_list')` rather than
+        // an empty Backbone Collection. Since this is a new feature, everything will be empty,
+        // so we can likely skip this improvement for a future refactor.
+        let activeAccessList = this.props.activeAccessList;
+        if (!activeAccessList) {
+            activeAccessList = new Backbone.Collection();
+        }
+
         return {
+            activeAccessList,
             visibility: this.props.visibility,
             users: this.props.users,
             imageUsers: this.props.imageUsers || new Backbone.Collection()
         }
     },
 
+    /** Callbacks */
     isSubmittable: function() {
         var hasVisibility = !!this.state.visibility;
         return hasVisibility;
@@ -41,6 +55,7 @@ export default React.createClass({
     onNext: function() {
         this.props.onNext({
             visibility: this.state.visibility,
+            activeAccessList: this.state.activeAccessList,
             imageUsers: this.state.imageUsers
         });
     },
@@ -48,15 +63,17 @@ export default React.createClass({
     onSubmit: function() {
         this.props.onSubmit({
             visibility: this.state.visibility,
+            activeAccessList: this.state.activeAccessList,
             imageUsers: this.state.imageUsers
         });
     },
 
     onVisibilityChange: function(newVisibility) {
-        // when we change visibility we should reset the user list to empty
+        // reset visibility will also reset the lists to empty
         this.setState({
             visibility: newVisibility,
-            imageUsers: new Backbone.Collection()
+            imageUsers: new Backbone.Collection(),
+            activeAccessList: new Backbone.Collection()
         });
     },
 
@@ -75,7 +92,32 @@ export default React.createClass({
             imageUsers: imageUsers
         })
     },
+    onPatternCreated: function(patternObj) {
+        let params = {
+            pattern: patternObj.pattern,
+            type: (patternObj.type == "E-Mail") ? 'Email': 'Username',
+            // Add a success callback to add new pattern to list
+            success: this.onAccessAdded
+        };
+        actions.PatternMatchActions.create(params);
+    },
 
+    onAccessAdded: function(pattern_match) {
+        let activeAccessList = this.state.activeAccessList
+        activeAccessList.add(pattern_match)
+        this.setState({
+            activeAccessList: activeAccessList
+        });
+    },
+    onAccessRemoved: function(pattern_match) {
+        let activeAccessList = this.state.activeAccessList
+        activeAccessList.remove(pattern_match)
+        this.setState({
+            activeAccessList: activeAccessList
+        });
+    },
+
+    /** Rendering **/
     renderUserList: function() {
         let helpLabel = "Please include users that should be able to launch this image.";
         if (this.state.visibility === "select") {
@@ -85,11 +127,32 @@ export default React.createClass({
         }
     },
 
+    renderAccessList: function() {
+        if (this.state.visibility === "select") {
+            if(this.props.allPatterns == null) {
+                return (<div className="loading" />);
+            }
+            return (
+            <div>
+                <hr />
+                <EditAccessView
+                    allPatterns={this.props.allPatterns}
+                    activeAccessList={this.state.activeAccessList}
+                    onAccessAdded={this.onAccessAdded}
+                    onAccessRemoved={this.onAccessRemoved}
+                    onCreateNewPattern={this.onPatternCreated}
+                />
+            </div>
+            );
+        }
+    },
+
     renderBody: function() {
         return (
         <div>
             <Visibility instance={this.props.instance} value={this.state.visibility} onChange={this.onVisibilityChange} />
             {this.renderUserList()}
+            {this.renderAccessList()}
         </div>
         );
     },
